+++ modules/item/item.js
@@ -1,4 +1,7 @@
 import { HMTABLES } from '../sys/constants.js';
+import { HMChatMgr } from '../mgr/chatmgr.js';
+import { HMDialogMgr } from '../mgr/dialogmgr.js';
+import { HMRollMgr } from '../mgr/rollmgr.js';
 
 // Remember: Items may not alter Actors under any circumstances.
 // You will create a free fire shooting gallery if you do this, and
@@ -321,6 +324,52 @@ export class HMItem extends Item {
 
         if (hp < 0) return this.delete();
         await this.update({'data': {hp, timer, treated}});
+    }
+
+    // TODO: Refactor. This is propagation of the mess in actor-sheet.js
+    static async rollSkill({skillName, specialty=null}) {
+        const actors = canvas.tokens.controlled.map((token) => token.actor);
+        const callers = [];
+        Object.values(actors).forEach(async (actor) => {
+            let context;
+            if (specialty) {
+                console.warn(specialty);
+            context = actor.items.find((a) => a.type === 'skill'
+                    && skillName === a.name
+                    && specialty === a.data.data?.specialty?.value);
+            } else {
+                context = actor.items.find((a) => a.type === 'skill'
+                    && skillName === a.name
+                    && !a.data.data?.specialty?.value);
+            }
+            if (!context) return;
+            callers.push({caller: actor, context});
+        });
+
+        if (!callers.length) return;
+
+        const dialogDataset = {
+            dialog: 'skill',
+            formula: 'd100 + @resp.mod @resp.oper @bonus.total.value',
+            skillType: 'skill',
+            itemId: callers[0].context.id,
+        };
+
+        const dialogMgr  = new HMDialogMgr();
+        const dialogResp = await dialogMgr.getDialog(dialogDataset, callers[0].caller);
+
+        Object.values(callers).forEach(async (caller) => {
+            const resp = caller;
+            resp.resp = dialogResp.resp;
+            const dataset = dialogDataset;
+            dataset.itemId = resp.context.id;
+
+            const rollMgr  = new HMRollMgr();
+            const roll     = await rollMgr.getRoll(dataset, resp);
+            const chatMgr  = new HMChatMgr();
+            const card     = await chatMgr.getCard(roll, dataset, resp);
+            await ChatMessage.create(card);
+        });
     }
 
     static specname(data) {