+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Skill macros.
 ### Changed
 - Switched canvas font to Gentium Book.
 - Added scroll bar to wound list.
+- Added to compendiums.
 
 ## [0.2.1] - 2022-04-13
 ### Fixed